# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: discogs-fade-backend
  version: 0.1.0
  description: >
    Boundary between the extension and Discogs' API. Wraps GET /masters/{id} and
    GET and POST /fade — master DTOs are currently unscoped (all Discogs response
    fields kept as-is). GET /releases/{id} serves a trimmed release from our
    own database, and GET /releases lists and filters the releases stored there.
paths:
  /releases:
    get:
      operationId: listReleases
      summary: List every enriched release
      description: >
        Pages, filters and sorts the stored releases; 400 on invalid input.
        Releases the user has faded (directly or through their master) are
        left out of items and total.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 50 }
        - $ref: "#/components/parameters/StyleCombo"
        - name: genre
          in: query
          description: Comma-separated genre names, OR-matched
          schema:
            type: string
        - name: style
          in: query
          description: Comma-separated style names, OR-matched
          schema:
            type: string
        - name: format
          in: query
          description: Comma-separated format names (e.g. Vinyl, CD), OR-matched
          schema:
            type: string
        - name: country
          in: query
          description: Comma-separated country names, OR-matched
          schema:
            type: string
        - name: tracks
          in: query
          description: Comma-separated track counts, OR-matched; each is 1 to 6, or 7+ for seven or more
          schema:
            type: string
        - name: excludeGenre
          in: query
          description: Comma-separated genre names; releases with any of them are left out
          schema:
            type: string
        - name: excludeStyle
          in: query
          description: Comma-separated style names; releases with any of them are left out
          schema:
            type: string
        - name: excludeFormat
          in: query
          description: Comma-separated format names; releases with any of them are left out
          schema:
            type: string
        - name: onlyGenre
          in: query
          description: Comma-separated genre names; keeps releases whose genres are all in the list
          schema:
            type: string
        - name: onlyStyle
          in: query
          description: Comma-separated style names; keeps releases whose styles are all in the list
          schema:
            type: string
        - name: onlyFormat
          in: query
          description: Comma-separated format names; keeps releases whose formats are all in the list
          schema:
            type: string
        - name: noLinks
          in: query
          description: "only keeps releases with no video links; exclude leaves them out"
          schema:
            type: string
            enum: [only, exclude]
        - name: yearMin
          in: query
          schema: { type: integer }
        - name: yearMax
          in: query
          schema: { type: integer }
        - name: sort
          in: query
          schema:
            type: string
            enum:
              [title, -title, year, -year, artist, -artist, format, -format, rating, -rating, tracks, -tracks]
            default: title
      responses:
        "200":
          description: Paginated releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseListPageDto"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /releases/facets:
    get:
      operationId: getReleaseFacets
      summary: >
        The genres, styles, format names and countries of all enriched releases,
        each with how many releases carry it, counted among the releases the same
        filters would list (faded ones excluded).
      parameters:
        - $ref: "#/components/parameters/Genre"
        - $ref: "#/components/parameters/Style"
        - $ref: "#/components/parameters/Format"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/Tracks"
        - $ref: "#/components/parameters/ExcludeGenre"
        - $ref: "#/components/parameters/ExcludeStyle"
        - $ref: "#/components/parameters/ExcludeFormat"
        - $ref: "#/components/parameters/OnlyGenre"
        - $ref: "#/components/parameters/OnlyStyle"
        - $ref: "#/components/parameters/OnlyFormat"
        - $ref: "#/components/parameters/NoLinks"
        - $ref: "#/components/parameters/YearMin"
        - $ref: "#/components/parameters/YearMax"
      responses:
        "200":
          description: Facets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseFacetsDto"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /releases/{id}:
    get:
      operationId: getRelease
      summary: >
        Get a release by id from our database; falls back to a live Discogs
        fetch (saved for next time) when it is not indexed yet.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDetailDto"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Release not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "502":
          description: Discogs fetch failed after retries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "503":
          description: Discogs is erroring and the queue is paused, retrying; try again shortly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "504":
          description: Timed out waiting for the Discogs fetch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /releases/{id}/refresh:
    post:
      operationId: refreshRelease
      summary: >
        Force-refresh a release: fetch it live from Discogs (ahead of any
        background indexing in the queue), overwrite everything stored for it,
        and return the fresh release.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDetailDto"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Release not found on Discogs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "502":
          description: Discogs fetch failed after retries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "503":
          description: Discogs is erroring and the queue is paused, retrying; try again shortly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "504":
          description: Timed out waiting for the Discogs fetch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
components:
  parameters:
    Genre:
      name: genre
      in: query
      description: Comma-separated genre names, OR-matched
      schema: { type: string }
    Style:
      name: style
      in: query
      description: Comma-separated style names, OR-matched
      schema: { type: string }
    Format:
      name: format
      in: query
      description: Comma-separated format names (e.g. Vinyl, CD), OR-matched
      schema: { type: string }
    Country:
      name: country
      in: query
      description: Comma-separated country names, OR-matched
      schema: { type: string }
    Tracks:
      name: tracks
      in: query
      description: Comma-separated track counts, OR-matched; each is 1 to 6, or 7+ for seven or more
      schema: { type: string }
    ExcludeGenre:
      name: excludeGenre
      in: query
      description: Comma-separated genre names; releases with any of them are left out
      schema: { type: string }
    ExcludeStyle:
      name: excludeStyle
      in: query
      description: Comma-separated style names; releases with any of them are left out
      schema: { type: string }
    ExcludeFormat:
      name: excludeFormat
      in: query
      description: Comma-separated format names; releases with any of them are left out
      schema: { type: string }
    OnlyGenre:
      name: onlyGenre
      in: query
      description: Comma-separated genre names; keeps releases whose genres are all in the list
      schema: { type: string }
    OnlyStyle:
      name: onlyStyle
      in: query
      description: Comma-separated style names; keeps releases whose styles are all in the list
      schema: { type: string }
    OnlyFormat:
      name: onlyFormat
      in: query
      description: Comma-separated format names; keeps releases whose formats are all in the list
      schema: { type: string }
    StyleCombo:
      name: styleCombo
      in: query
      description: >
        A style combination: comma-separated style names a release must all have.
        Repeat the parameter for several combinations; a release passes when it
        matches any one of them. `GET /releases` only; the facets ignore it, so
        counts do not follow combinations.
      style: form
      explode: true
      schema:
        type: array
        items: { type: string }
    NoLinks:
      name: noLinks
      in: query
      description: "only keeps releases with no video links; exclude leaves them out"
      schema:
        type: string
        enum: [only, exclude]
    YearMin:
      name: yearMin
      in: query
      schema: { type: integer }
    YearMax:
      name: yearMax
      in: query
      schema: { type: integer }
  schemas:
    ErrorDto:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ReleaseFormatDto:
      type: object
      required: [name, descriptions]
      properties:
        name: { type: string }
        descriptions:
          type: array
          items: { type: string }
    ReleaseArtistDto:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    ReleaseListItemDto:
      type: object
      required:
        [releaseId, title, thumb, year, country, genres, styles, formats, ratingAverage, ratingCount, haves, wants, artists]
      properties:
        releaseId: { type: integer }
        title: { type: string }
        thumb: { type: string, nullable: true }
        year: { type: integer, nullable: true }
        country: { type: string, nullable: true }
        genres:
          type: array
          items: { type: string }
        styles:
          type: array
          items: { type: string }
        formats:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseFormatDto"
        ratingAverage: { type: number, nullable: true }
        ratingCount: { type: integer, nullable: true }
        haves: { type: integer, nullable: true }
        wants: { type: integer, nullable: true }
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseArtistDto"
    ReleaseListPageDto:
      type: object
      required: [items, page, pageSize, total, enrichedCount, fadedCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseListItemDto"
        page: { type: integer }
        pageSize: { type: integer }
        total:
          type: integer
          description: Releases matching the current filters, faded ones excluded.
        enrichedCount:
          type: integer
          description: All enriched releases, faded ones included. Ignores filters.
        fadedCount:
          type: integer
          description: How many of enrichedCount are faded for the user. Ignores filters.
    FacetValueDto:
      type: object
      required: [value, count]
      properties:
        value: { type: string }
        count:
          type: integer
          description: How many of the releases matching the filters carry the value.
    ReleaseFacetsDto:
      type: object
      description: >
        The values of each category among the releases matching the filters, most
        common first (ties alphabetical). Values no matching release carries are left out.
      required: [genres, styles, formats, countries]
      properties:
        genres:
          type: array
          items: { $ref: "#/components/schemas/FacetValueDto" }
        styles:
          type: array
          items: { $ref: "#/components/schemas/FacetValueDto" }
        formats:
          type: array
          items: { $ref: "#/components/schemas/FacetValueDto" }
        countries:
          type: array
          items: { $ref: "#/components/schemas/FacetValueDto" }
...
    ReleaseDetailDto:
      type: object
      description: >
        Served from our own database (fetched live from Discogs and saved on
        first request if the release is not indexed yet).
      required: [id, title, year, country, artists, tracklist, videos]
      properties:
        id: { type: integer }
        title: { type: string }
        thumb: { type: string }
        year: { type: integer, nullable: true }
        country: { type: string, nullable: true }
        genres:
          type: array
          items: { type: string }
        styles:
          type: array
          items: { type: string }
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseArtistStub"
        tracklist:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseTrack"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseVideo"
    ReleaseArtistStub:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    ReleaseTrack:
      type: object
      required: [position, title]
      properties:
        position: { type: string }
        title: { type: string }
        duration: { type: string }
    ReleaseVideo:
      type: object
      required: [uri]
      properties:
        uri: { type: string }
        title: { type: string }
        duration: { type: integer }
...